Fix: 헤더 기능 수정
로그인전과 후로 나뉘어져있던 헤더 v-if문을 활용해서 하나로 합치고 라우터 설정 및 app.vue css 수정
