+++ src/router/router.js
@@ -13,6 +13,10 @@ const router = createRouter({
             component: () => import('../views/LoginPage.vue')
         },
         {
+            path: '/after/login',
+            component: () => import('../views/MainLogin.vue')
+        },
+        {
             path: '/regist',
             component: () => import('../views/Regist.vue')
         },
@@ -43,6 +47,22 @@ const router = createRouter({
         {
             path: '/recruit-modify/:id',
             component: () => import('@/components/Recruit/RecruitModify.vue')
+        },
+        {
+            path: '/findId',
+            component: () => import('../views/FindIdPage.vue')
+        },
+        {
+            path: '/findIdResult/:email',
+            component: () => import('../views/FindIdResultPage.vue')
+        },
+        {
+            path: '/resetPassword',
+            component: () => import('../views/ResetPwdPage.vue')
+        },
+        {
+            path: '/resetPasswordFinal/:email',
+            component: () => import('../views/ResetPwdFinalPage.vue')
         }
     ]
 })
